+++ src/prpl.ts
@@ -15,6 +15,7 @@
 import * as capabilities from 'browser-capabilities';
 import * as fs from 'fs';
 import * as http from 'http';
+import * as httpErrors from 'http-errors';
 import * as path from 'path';
 import * as send from 'send';
 import * as url from 'url';
@@ -25,6 +26,14 @@ export interface Config {
   // The Cache-Control header to send for all requests except the entrypoint.
   // Defaults to `max-age=60`.
   cacheControl?: string;
+
+  // A custom error-handling logic function.
+  //
+  // This function could be useful if you want, by example, be able to
+  // render custom errors pages.
+  error?:
+      (request: http.IncomingMessage,
+       response: http.ServerResponse, error: httpErrors.HttpError) => void;
 
   // Serves a tiny self-unregistering service worker for any request path
   // ending with `service-worker.js` that would otherwise have had a 404 Not
@@ -84,8 +93,12 @@ export function makeHandler(root?: string, config?: Config): (
     // is a prefix of "/foo-secrets".
     const absFilepath = path.normalize(path.join(absRoot, urlPath));
     if (!absFilepath.startsWith(addTrailingPathSep(absRoot))) {
-      response.writeHead(403);
-      response.end('Forbidden');
+      if (config && config.error) {
+        config.error(request, response, httpErrors(403));
+      } else {
+        response.writeHead(403);
+        response.end('Forbidden');
+      }
       return;
     }
 
@@ -107,8 +120,12 @@ export function makeHandler(root?: string, config?: Config): (
     // that we only return this error for the entrypoint; we always serve fully
     // qualified static resources.
     if (!build && serveEntrypoint) {
-      response.writeHead(500);
-      response.end('This browser is not supported.');
+      if (config && config.error) {
+        config.error(request, response, httpErrors(500));
+      } else {
+        response.writeHead(500);
+        response.end('This browser is not supported.');
+      }
       return;
     }
 
@@ -156,7 +173,16 @@ self.addEventListener('activate', () => self.registration.unregister());`);
       // We handle the caching header ourselves.
       cacheControl: false,
     };
-    send(request, fileToSend, sendOpts).pipe(response);
+
+    let stream = send(request, fileToSend, sendOpts);
+
+    // Set a custom error-handling function
+    if (config && config.error) {
+      const errorFunction = (error:  httpErrors.HttpError) => config.error!(request, response, error)
+      stream = stream.on('error', errorFunction);
+    }
+
+    stream.pipe(response);
   };
 }