+++ src/test/prpl_test.ts
@@ -14,6 +14,7 @@
 
 import * as capabilities from 'browser-capabilities';
 import {assert} from 'chai';
+import * as express from 'express';
 import * as http from 'http';
 import * as httpErrors from 'http-errors';
 import * as path from 'path';
@@ -28,7 +29,7 @@ suite('prpl server', function() {
   let host: string;
   let port: number;
 
-  const startServer = (root: string, config?: prpl.Config): Promise<void> => {
+  function startServer(root: string, config?: prpl.Config): Promise<void> {
     const handler = prpl.makeHandler(root, config);
     server = http.createServer(
         (request: http.IncomingMessage, response: http.ServerResponse) => {
@@ -48,14 +49,14 @@ suite('prpl server', function() {
         resolve();
       });
     });
-  };
+  }
 
-  const get =
-      (path: string, ua?: string, headers?: {[key: string]: string}): Promise<{
-        code: number | undefined,
-        data: string,
-        headers: {[key: string]: string}
-      }> => {
+  type GetResponse = {
+    code: number|undefined; data: string; headers: {[key: string]: string};
+  }
+
+  function get(path: string, ua?: string, headers?: {[key: string]: string}):
+      Promise<GetResponse> {
         return new Promise((resolve) => {
           const getHeaders = Object.assign({'user-agent': ua || ''}, headers);
           http.get({host, port, path, headers: getHeaders}, (response) => {
@@ -66,7 +67,15 @@ suite('prpl server', function() {
             response.on('end', () => resolve({code, data, headers}));
           });
         });
-      };
+      }
+
+  function checkPlainTextError(
+      expectCode: number, expectData: string, res: GetResponse) {
+    assert.equal(res.code, expectCode);
+    assert.equal(res.data, expectData);
+    assert.equal(res.headers['content-type'], 'text/plain');
+    assert.equal(res.headers['content-length'], String(expectData.length));
+  }
 
   suite('configured with multiple builds', () => {
     suiteSetup(async () => {
@@ -107,23 +116,18 @@ suite('prpl server', function() {
       });
 
       test('serves a 404 for missing file with extension', async () => {
-        const {code} = await get('/foo.png');
-        assert.equal(code, 404);
+        checkPlainTextError(404, 'Not Found', await get('/foo.png'));
       });
 
       test('forbids traversal outside root', async () => {
-        const {code, data} = await get('/../secrets');
-        assert.equal(code, 403);
-        assert.equal(data, 'Forbidden');
+        checkPlainTextError(403, 'Forbidden', await get('/../secrets'));
       });
 
       test('forbids traversal outside root with matching prefix', async () => {
         // Edge case where the resolved request path naively matches the root
         // directory by prefix even though it's actually a sibling, not a child
         // ("/static-secrets" begins with "/static").
-        const {code, data} = await get('/../static-secrets');
-        assert.equal(code, 403);
-        assert.equal(data, 'Forbidden');
+        checkPlainTextError(403, 'Forbidden', await get('/../static-secrets'));
       });
     });
 
@@ -147,8 +151,7 @@ suite('prpl server', function() {
       });
 
       test('serves a 404 for missing file with extension', async () => {
-        const {code} = await get('/foo.png', chrome);
-        assert.equal(code, 404);
+        checkPlainTextError(404, 'Not Found', await get('/foo.png'));
       });
 
       test('sets push headers for fragment', async () => {
@@ -238,9 +241,8 @@ suite('prpl server', function() {
     });
 
     test('serves 500 error to unsupported browser', async () => {
-      const {code, data} = await get('/');
-      assert.equal(code, 500);
-      assert.include(data, 'not supported');
+      checkPlainTextError(
+          500, 'This browser is not supported.', await get('/'));
     });
   });
 
@@ -295,59 +297,54 @@ suite('prpl server', function() {
     });
   });
 
-  suite('configured with an error function', () => {
-    suiteSetup(async () => {
-      await startServer(path.join('src', 'test', 'static'), {
+  suite('configured with express error forwarding', () => {
+    suiteSetup((done) => {
+      const app = express();
+
+      app.use(prpl.makeHandler(path.join('src', 'test', 'static'), {
+        forwardErrors: true,
         builds: [
           {
             name: 'es2015',
             browserCapabilities: ['es2015' as capabilities.BrowserCapability],
           },
-        ],
-        error: (req: http.IncomingMessage, res: http.ServerResponse, err: httpErrors.HttpError) => {
-              res.statusCode = err.status || 500
-              const customErrorPages = [403, 404, 500];
-              const hasCustomErrorPage =
-                  customErrorPages.includes(res.statusCode)
+        ]
+      }));
 
-              if (hasCustomErrorPage) {
-                res.end(`Custom ${res.statusCode} error page for ${req.url}`);
-              }
-              else {
-                res.end(err.message)
-              }
-            }
+      app.use(
+          (error: httpErrors.HttpError,
+           _request: any,
+           response: any,
+           _next: express.NextFunction) => {
+            response.statusCode = error.status;
+            response.setHeader('Content-Type', 'text/plain');
+            response.end(`custom ${error.status}: ${error.message}`);
+          });
+
+      server = app.listen(/* random */ 0, '127.0.0.1', () => {
+        host = server.address().address;
+        port = server.address().port;
+        done();
       });
     });
 
-    test(
-        'should render a custom 403 error page for directory traversal attack',
-        async () => {
-          const status = 403;
-          const url = '/../secrets';
-          const {code, data} = await get(url);
-          assert.equal(code, status);
-          assert.equal(data, `Custom ${status} error page for ${url}`);
-        });
+    suiteTeardown((done) => {
+      server.close(done);
+    });
 
-    test(
-        'should return a custom 404 error page for a Not Found file',
-        async () => {
-          const status = 404;
-          const url = '/fragment/error.html';
-          const {code, data} = await get(url);
-          assert.equal(code, status);
-          assert.include(data, `Custom ${status} error page for ${url}`);
-        });
+    test('forwards error for 404 not found', async () => {
+      checkPlainTextError(
+          404, 'custom 404: Not Found', await get('/fragment/error.html'));
+    });
 
-    test(
-        'should render a custom 500 error page to unsupported browsers',
-        async () => {
-          const status = 500;
-          const url = '/';
-          const {code, data} = await get(url);
-          assert.equal(code, status);
-          assert.equal(data, `Custom ${status} error page for ${url}`);
-        });
+    test('forwards error for directory traversal 403', async () => {
+      checkPlainTextError(
+          403, 'custom 403: Forbidden', await get('/../secrets'));
+    });
+
+    test('forwards error for unsupported browser 500', async () => {
+      checkPlainTextError(
+          500, 'custom 500: This browser is not supported.', await get('/'));
+    });
   });
 });