+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Add `forwardErrors` option to pass 404s and other HTTP errors down to the next Express error-handling middleware.
+- Recommend `npm` instead of `yarn` and switch to `npm` lock file.
 
 ## [0.11.0] 2017-10-23
 - Add `unregisterMissingServiceWorkers` option (default true) which serves a tiny self-unregistering service worker for would-be 404 service worker requests, to prevent clients from getting stuck with invalid service workers indefinitely.